% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{cell_to_children_size}
\alias{cell_to_children_size}
\title{Cell to children size}
\usage{
cell_to_children_size(h3_address = NULL, child_res = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character index generated by H3.}

\item{child_res}{numeric; child resolution to report on.}

\item{simple}{Logical; whether to return a vector or a data frame containing
both inputs and outputs.}
}
\value{
numeric; number of children at the requested resolution
}
\description{
Get the number of children for a cell at a given resolution.
}
\examples{
# example address has resolution 7:
cell_to_children_size('872830b82ffffff', c(8,9,10,11), simple = FALSE)
}
