% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{degs_to_rads}
\alias{degs_to_rads}
\title{Convert degrees to radians}
\usage{
degs_to_rads(degree = NULL, lang = c("r", "h3"), simple = TRUE)
}
\arguments{
\item{degree}{Numeric, value in degrees}

\item{lang}{Character; whether to perform the conversion using base R
or the H3 library. Defaults to R for speed.}

\item{simple}{Logical; whether to return a vector or a data frame containing
both inputs and outputs.}
}
\value{
Numeric, value in radians
}
\description{
Convert degrees to radians.
}
\examples{
degs_to_rads(120)
}
