% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{edge_length}
\alias{edge_length}
\title{H3 edge length (exact)}
\usage{
edge_length(h3_address = NULL, units = c("m", "km", "rads"), simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 16-character index generated by H3.}

\item{units}{Length unit to report in, either meters or kilometers.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a numeric vector of length(h3_address).
}
\description{
This function calculates the exact length of an H3 edge.
}
\examples{
edge_length(h3_address = '166be8d12fffffff', 'm')
}
