% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{geo_to_h3}
\alias{geo_to_h3}
\title{Convert point location to H3 address}
\usage{
geo_to_h3(lon = NULL, lat = NULL, res = NULL)
}
\arguments{
\item{lon}{Number; Longitude in decimal degrees and WGS84 datum.}

\item{lat}{Number; Latitude in decimal degrees and WGS84 datum}

\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}
}
\value{
A dataframe with input coordinates, resolution, and H3 address as
  columns.
}
\description{
This function takes a latitude and longitude in WGS84 and returns a H3
address at the chosen resolution.
}
\note{
You can supply multiple coordinates and resolutions as vectors, but
  they must be recyclable against each other as they are coerced into a data
  frame. It is safest to provide many points and one resolution, or one point
  and many resolutions.
}
\examples{
# where is the Brisbane Town Hall at resolution 15?
brisbane <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = 15)

# where is it at multiple resolutions?
brisbane_all <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = seq(15))
}
