% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{get_cell_vertexes}
\alias{get_cell_vertexes}
\title{Get all vertex indexes}
\usage{
get_cell_vertexes(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character cell index generated by H3. A
vector of indexes can also be supplied.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a list of \code{length(h3_address)}.
}
\description{
This function returns all 6 vertex indices for a supplied H3 cell.
}
\examples{
# Get vertexes for this cell
get_cell_vertexes(h3_address = '86be8d12fffffff')

}
