% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_coordinates.R
\name{get_local_h3}
\alias{get_local_h3}
\title{Get H3 address from local i, j coordinates}
\usage{
get_local_h3(origin = NULL, coords = NULL, simple = TRUE)
}
\arguments{
\item{origin}{Character; 15-character address generated by H3. A vector of
addresses can also be supplied.}

\item{coords}{List with 2 integer elements; a single i coordinate or vector
of same, and a single j coordinate or vector of same. List elements must
have the same length.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
If `simple = TRUE`, a character vector of H3 addresses. If not, a
  data frame containing columns origin, i, j, destination.
}
\description{
This function returns H3 destination addresses for local i, j coordinate
pairs anchored by an H3 origin address.
}
\note{
\itemize{
  \item{The coordinate space used by this function may have deleted regions
  or warping due to pentagonal distortion.}
  \item{Coordinates are only comparable if they come from the same origin
  index.}
  \item{Failure may occur if the index is too far away from the origin or if
  the index is on the other side of a pentagon.}
  \item{This function is experimental, and its output is not guaranteed to be
  compatible across different versions of H3.}
 }
}
\examples{
# Get local coordinates for a nearby hexagon
local <- get_local_ij(origin = '86be8d12fffffff', destination = '86be8d127ffffff')

# Convert back to destination address
get_local_h3(origin = local$origin, coords = list(local$local_i, local$local_j))

}
