% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{get_res0}
\alias{get_res0}
\title{Get resolution 0 indexes}
\usage{
get_res0()
}
\value{
length 122 character vector of top-level H3 cell indices.
}
\description{
Get all H3 cell indexes at resolution 0.
}
\note{
As every index at every resolution > 0 is
the descendant of a res 0 index, this can be used with
\code{\link[h3jsr:get_children]{get_children}} to iterate over H3 indexes at
any resolution.
}
\examples{
res0 <- get_res0()
cell_area(res0[1], 'km2')
}
