% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_get_base_cell}
\alias{h3_get_base_cell}
\title{get the base cell of an H3 address}
\usage{
h3_get_base_cell(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an integer vector of length(h3_address), ranging from
  0 to 121.
}
\description{
This function returns the number of the base (Level 1) cell for an H3
address.
}
\examples{
# What is Brisbane Town Hall's base cell number?
h3_get_base_cell(h3_address = '8abe8d12acaffff')
}
