% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_get_res}
\alias{h3_get_res}
\title{get the resolution of an H3 address}
\usage{
h3_get_res(h3_address = NULL)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}
}
\value{
A dataframe with input H3 address(es) and Integer column containing
  resolution number, 1-15.
}
\description{
This function returns an H3 address' resolution level.
}
\examples{
# What is the H3 resolution of this address?
h3_get_res(h3_address = '8abe8d12acaffff')
}
