% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{h3_get_ring}
\alias{h3_get_ring}
\title{get a donut of H3 addresses}
\usage{
h3_get_ring(h3_address = NULL, ring_size = 1, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{ring_size}{Character; number of steps away from the central address.
Defaults to 1.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a list of length(h3_address). Each list element contains
  a character vector of H3 addresses belonging to that step away from the
  input address.
}
\description{
This function returns all the H3 addresses at the specified step from the
address supplied.
}
\note{
In total, the number of addresses returned for each input address is
  `ring_size \* 6`.
}
\examples{
# What are the neighbours of this address at step 2?
h3_get_ring(h3_address = '86be8d12fffffff', ring_size = 2)
}
