% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{h3_get_udedges}
\alias{h3_get_udedges}
\title{get all unidirectional edges}
\usage{
h3_get_udedges(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, list of length(h3_address). Each list contains a
  character vector of H3 edge addresses.
}
\description{
Get all unidirectional edges for a given H3 address.
}
\examples{
# Get all the edges for this address
h3_get_udedges(h3_address = '86be8d12fffffff')

}
