% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_is_pentagon}
\alias{h3_is_pentagon}
\title{check if H3 address is a pentagon}
\usage{
h3_is_pentagon(h3_address = NULL)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}
}
\value{
A dataframe with input H3 address and Boolean column (TRUE if address
  is a pentagon).
}
\description{
This function checks whether a H3 address refers to one of the pentagons
that occur at icosahedron corners.
}
\examples{
# is the following address a pentagon?
h3_is_pentagon(h3_address = '8abe8d12acaffff')
}
