% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_is_valid}
\alias{h3_is_valid}
\title{check H3 address}
\usage{
h3_is_valid(h3_address = NULL, simple = FALSE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a data frame or logical vector of
outputs.}
}
\value{
A dataframe with input H3 address and Boolean column (TRUE if address
  is valid). If `simple = TRUE``, a logical vector of length(h3_address).
}
\description{
This function checks whether an H3 address is valid.
}
\examples{
# is the following address valid?
h3_is_valid(h3_address = '8abe8d12acaffff')
}
