% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{h3_res_area}
\alias{h3_res_area}
\title{get h3 address area}
\usage{
h3_res_area(res = NULL, units = c("m2", "km2"), fast = TRUE)
}
\arguments{
\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{units}{Areal unit to report in, either square meters or square
kilometers.}

\item{fast}{Logical; whether to retieve values from a locally stored table or
reclaculate from source.}
}
\value{
Numeric; average h3 address area.
}
\description{
This function returns the average area of an H3 address at a given
resolution.
}
\examples{
# Return average h3 address area at each resolution in square meters
h3_res_area(res = seq(0, 15), units = 'm2')

}
