% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{h3_res_edgelen}
\alias{h3_res_edgelen}
\title{get h3 address edge length}
\usage{
h3_res_edgelen(res = NULL, units = c("m", "km"), fast = TRUE)
}
\arguments{
\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{units}{Length unit to report in, either meters or kilometers.}

\item{fast}{Logical; whether to retieve values from a locally stored table or
reclaculate from source.}
}
\value{
Numeric; h3 address edge length
}
\description{
This function returns the average edge length of an H3 address at a given
resolution.
}
\examples{
# Return average h3 address edge length at each resolution in kilometers
h3_res_edgelen(res = seq(0, 15), units = 'km')

}
