% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{get the bounding points of a H3 address}
\usage{
h3_to_geo_boundary(h3_address = NULL)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}
}
\value{
A dataframe with two columns. One contains the input h3 address and
  the other is a list-column of matrices suitable for use with
  `sf::st_polygon()`.
}
\description{
This function takes a H3 address and returns the coordinates of its bounding
shape (usually a hexagon) in WGS84.
}
\examples{
# What is the hexagon over the Brisbane Town Hall at resolution 10?
brisbane_hex_10 <- h3_to_geo_boundary(h3_address = '8abe8d12acaffff')

}
