% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{h3_to_geo_udedge}
\alias{h3_to_geo_udedge}
\title{get the geometry of an H3 edge}
\usage{
h3_to_geo_udedge(h3_edge = NULL, simple = TRUE)
}
\arguments{
\item{h3_edge}{Address of unidirectional edge.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an object of type `sfc_LINESTRING`.
}
\description{
This function takes an H3 edge address and returns the coordinates of its
geometry in WGS84.
}
\examples{
# get me the shape of this edge
h3_to_geo_udedge(h3_edge = '166be8d12fffffff')

}
