% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{h3_uncompact}
\alias{h3_uncompact}
\title{uncompact H3 geometry}
\usage{
h3_uncompact(h3_addresses = NULL, res = NULL, simple = TRUE)
}
\arguments{
\item{h3_addresses}{Character vector or list of 15-character addresses
generated by H3 at a single resolution, generally the output of
`h3_polyfill()`.}

\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{simple}{Logical; whether to return a vector of outputs or a list object
containing both inputs and outputs.}
}
\value{
A list of H3 addresses of the chosen resolution.
}
\description{
This function uncompacta a compacted set of hexagons to hexagons of the same
resolution.
}
\examples{
\dontrun{
# Give me a compacted representation of County Ashe, NC
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
nc1 <- nc[1, ]
nc1 <- sf::st_cast(nc1, 'POLYGON')
fillers <- h3_polyfill(geometry = nc1, res = 6)
compacted <- h3_compact(fillers)
# uncompact to resolution 7
uncompacted <- h3_uncompact(compacted, res = 7)
}
}
