% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{is_pentagon}
\alias{is_pentagon}
\title{check if H3 cell index is a pentagon}
\usage{
is_pentagon(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character index generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a logical vector of \code{length(h3_address)}.
}
\description{
This function checks whether a H3 cell index refers to one of the pentagons
that occur at icosahedron corners.
}
\examples{
# is the following cell index a pentagon?
is_pentagon(h3_address = '8abe8d12acaffff')
}
