% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{is_valid}
\alias{is_valid}
\title{check H3 address}
\usage{
is_valid(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a logical vector of length(h3_address).
}
\description{
This function checks whether an H3 address is valid.
}
\examples{
# is the following address valid?
is_valid(h3_address = '8abe8d12acaffff')
}
