% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertexes.R
\name{is_valid_vertex}
\alias{is_valid_vertex}
\title{check H3 cell index}
\usage{
is_valid_vertex(h3_vertex = NULL, simple = TRUE)
}
\arguments{
\item{h3_vertex}{Character; 15-character index generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a logical vector of \code{length(h3_vertex)}.
}
\description{
This function checks whether an H3 cell index is valid.
}
\examples{
# is the following cell index valid?
is_valid_vertex(h3_vertex = '25abe8d12ac87fff')
}
