% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{min_path}
\alias{min_path}
\title{find a path between H3 addresses}
\usage{
min_path(origin = NULL, destination = NULL, simple = TRUE)
}
\arguments{
\item{origin}{Character; 15-character address generated by H3. A vector of
addresses can also be supplied.}

\item{destination}{Character; 15-character address generated by H3. A vector
of addresses can also be supplied.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
A list of H3 addresses if simple = TRUE, an sf LINESTRING if not.
}
\description{
This function finds a minimum-steps path between two H3 addresses of the same
resolution.
}
\note{
\itemize{
  \item{The number of addresses supplied to origin and
destination must be equal.}
  \item{Multiple minimum-step pathways exist between any two hexagons; this
  function will pick one at random.}
 }
}
\examples{
# Draw a path between Brisbane and Sydney at resolution 6
min_path(origin = '86be8d127ffffff', destination = '86be0e35fffffff', simple = FALSE)

}
