% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_for_pt2cell}
\alias{prep_for_pt2cell}
\alias{prep_for_pt2cell.sf}
\alias{prep_for_pt2cell.sfc}
\alias{prep_for_pt2cell.sfg}
\alias{prep_for_pt2cell.matrix}
\alias{prep_for_pt2cell.data.frame}
\alias{prep_for_pt2cell.numeric}
\title{Prepare inputs for point_to_cell}
\usage{
prep_for_pt2cell(input = NULL)

\method{prep_for_pt2cell}{sf}(input = NULL)

\method{prep_for_pt2cell}{sfc}(input = NULL)

\method{prep_for_pt2cell}{sfg}(input = NULL)

\method{prep_for_pt2cell}{matrix}(input = NULL)

\method{prep_for_pt2cell}{data.frame}(input = NULL)

\method{prep_for_pt2cell}{numeric}(input = NULL)
}
\arguments{
\item{input}{\code{sf}, \code{sfc} or \code{sfg} POINT/MULTIPOINT object,
data frame or matrix. Data frames or matrices must have x, y coordinates in
their first two columns. WGS84 input is assumed in all cases.}
}
\value{
matrix representation of supplied coordinates.
}
\description{
Sets up a variety of possible input objects for
\code{\link[h3jsr:point_to_cell]{h3jsr::point_to_cell()}}.
}
\keyword{internal}
