% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{rads_to_degs}
\alias{rads_to_degs}
\title{Convert radians to degrees}
\usage{
rads_to_degs(radian = NULL, lang = c("r", "h3"), simple = TRUE)
}
\arguments{
\item{radian}{Numeric, value in radians}

\item{lang}{Character; whether to perform the conversion using base R
or the H3 library. Defaults to R for speed.}

\item{simple}{Logical; whether to return a vector or a data frame containing
both inputs and outputs.}
}
\value{
Numeric, value in degrees
}
\description{
Convert radians to degrees.
}
\examples{
rads_to_degs(1.5)
}
