% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{res_count}
\alias{res_count}
\title{Get total H3 addresses}
\usage{
res_count(res = NULL, fast = TRUE)
}
\arguments{
\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{fast}{Logical; whether to retieve values from a locally stored table or
reclaculate from source.}
}
\value{
Numeric; h3 address count.
}
\description{
This function returns total number of h3 addresses at a given resolution.
}
\examples{
# Return h3 address count for resolution 8
res_count(res = 8)

}
