% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{udedge_to_line}
\alias{udedge_to_line}
\title{Get the geometry of an H3 edge}
\usage{
udedge_to_line(h3_edge = NULL, simple = TRUE)
}
\arguments{
\item{h3_edge}{Address of directed edge.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an object of type `sfc_LINESTRING`.
}
\description{
This function takes an H3 unidirectional edge address and returns the
coordinates of its geometry in WGS84.
}
\examples{
# get me the shape of this edge
udedge_to_line(h3_edge = '166be8d12fffffff')

}
