% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{edge_length}
\alias{edge_length}
\title{Get exact cell edge length}
\usage{
edge_length(h3_edge = NULL, units = c("m", "km", "rads"), simple = TRUE)
}
\arguments{
\item{h3_edge}{Character; address of unidirectional edge.}

\item{units}{Length unit to report in. Options are meters, kilometers, or
radians.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a numeric vector of length(h3_address).
}
\description{
This function calculates the exact length of an H3 cell edge.
}
\examples{
edge_length(h3_edge = '166be8d12fffffff', 'm')
}
