% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{get_uddest}
\alias{get_uddest}
\title{Get destination from unidirectional edge}
\usage{
get_uddest(h3_edge = NULL, simple = TRUE)
}
\arguments{
\item{h3_edge}{Address of unidirectional edge.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, character vector of h3 addresses.
}
\description{
Get an H3 address representing the destination of a unidirectional edge.
}
\examples{
# Get the destination of this edge
get_uddest(h3_edge = '166be8d12fffffff')

}
