% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{get_udedge}
\alias{get_udedge}
\title{Get a unidirectional edge}
\usage{
get_udedge(origin = NULL, destination = NULL, simple = TRUE)
}
\arguments{
\item{origin}{Character; 15-character address generated by H3. A vector of
addresses can also be supplied.}

\item{destination}{Character; 15-character address generated by H3. A vector
of addresses can also be supplied.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, character vector of edges.
}
\description{
Get an H3 address representing a unidirectional edge for a given origin and
destination.
}
\note{
The number of addresses supplied to origin and destination must be
  equal.
}
\examples{
# Get me the edge between these two addresses
get_udedge(origin = '86be8d12fffffff', destination = '86be8d127ffffff')

}
