% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidirectional_edges.R
\name{get_udends}
\alias{get_udends}
\title{Get origin and destination of unidirectional edge}
\usage{
get_udends(h3_edge = NULL, simple = TRUE)
}
\arguments{
\item{h3_edge}{Address of unidirectional edge.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, character matrix of h3 addresses.
}
\description{
Get an H3 addresses representing the origin and destination of a
unidirectional edge.
}
\examples{
# Get the origin and destination of this edge
get_udends(h3_edge = '166be8d12fffffff')

}
