% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_get_base_cell}
\alias{h3_get_base_cell}
\title{get the base cell of an H3 address}
\usage{
h3_get_base_cell(h3_address = NULL)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}
}
\value{
A dataframe with input H3 address(es) and Integer column containing
  index number of base cell(s), 0-121.
}
\description{
This function returns the number of the base (Level 1) cell for an H3
address.
}
\examples{
# What is Brisbane Town Hall's base cell number?
h3_get_base_cell(h3_address = '8abe8d12acaffff')
}
