% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_is_rc3}
\alias{h3_is_rc3}
\title{check if H3 address is a in a Class III resolution}
\usage{
h3_is_rc3(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, a logical vector of length(h3_address).
}
\description{
This function checks whether a H3 address is in a Class III resolution
(rotated versus the icosahedron and subject to shape distortion adding extra
points on icosahedron edges, making them not true hexagons).
}
\examples{
# is the following address Class III?
h3_is_rc3(h3_address = '8abe8d12acaffff')
}
