% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{h3_set_to_multipolygon}
\alias{h3_set_to_multipolygon}
\title{get geometry for a set of H3 addresses}
\usage{
h3_set_to_multipolygon(h3_addresses = NULL, simple = TRUE)
}
\arguments{
\item{h3_addresses}{Character vector or list of 15-character addresses generated by
H3. Must}

\item{simple}{Logical; whether to return a single multipolygon or cast the
output to polygon.}
}
\value{
By default, object of type sfc_MULTIPOLYGON of length 1.
}
\description{
This function returns geometry associated with a set of H3 addresses, as a
single sfc_MULTIPOLYGON.
}
\note{
The geometry returned by this function will not be valid where the
  addresses supplied overlap. It is safest to restrict inputs to the results
  of, e.g., a single `h3_get_kring()` call.
}
\examples{
# Give me the outline of the hexagons around Brisbane Town Hall at
# resolution 10
bth <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = 10)
bth_patch <- h3_get_kring(h3_address = bth, ring_size = 2)
bth_patch_sf <- h3_set_to_multipolygon(bth)

}
