% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_to_geo}
\alias{h3_to_geo}
\title{Convert H3 address to point location}
\usage{
h3_to_geo(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an `sfc_POINT` object of `length(h3_address)`.
  EPSG:WGS84.
}
\description{
This function takes a H3 address and returns the coordinates of the center of
that H3 hexagon in WGS84.
}
\examples{
# Where is the center of the hexagon over the Brisbane Town Hall at resolution 10?
brisbane_10 <- h3_to_geo(h3_address = '8abe8d12acaffff')

}
