% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{get the bounding points of a H3 address}
\usage{
h3_to_geo_boundary(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an object of type `sfc_POLYGON`.
}
\description{
This function takes a H3 address and returns its bounding shape (usually a
hexagon) in WGS84.
}
\examples{
# What is the hexagon over the Brisbane Town Hall at resolution 10?
brisbane_hex_10 <- h3_to_geo_boundary(h3_address = '8abe8d12acaffff')

# Give me some of the hexes over Brisbane Town Hall as an sf object
BTH_hexes <-
  h3_to_geo_boundary(h3_address = geo_to_h3(lon = 153.023503,
                                            lat = -27.468920,
                                            res = seq(10, 15),
                                            simple = FALSE))
}
