% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{Get the boundary of an H3 address}
\usage{
h3_to_geo_boundary(h3_address = NULL, simple = TRUE)
}
\arguments{
\item{h3_address}{Character; 15-character address generated by H3.}

\item{simple}{Logical; whether to return a vector of outputs or a data frame
containing both inputs and outputs.}
}
\value{
By default, an `sfc_POLYGON` object of `length(h3_address)`.
}
\description{
This function takes an H3 address and returns its bounding shape (usually a
hexagon) in WGS84.
}
\examples{
# What is the hexagon over the Brisbane Town Hall at resolution 10?
brisbane_hex_10 <- h3_to_geo_boundary(h3_address = '8abe8d12acaffff')

# Give me some of the hexes over Brisbane Town Hall as an sf object
bth <- sf::st_sfc(sf::st_point(c(153.023503, -27.468920)), crs = 4326)
bth_addys <- unlist(geo_to_h3(bth, res = seq(10, 15)), use.names = FALSE)
bth_hexes <- h3_to_geo_boundary(h3_address = bth_addys)
plot(bth_hexes, axes = TRUE)
}
