% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3jsr.R
\docType{package}
\name{h3jsr-package}
\alias{h3jsr}
\alias{h3jsr-package}
\title{h3jsr: Access Uber's H3 library}
\description{
This package uses [`V8`](https://github.com/jeroen/v8) to access the
[javascript bindings for Uber's H3 library](https://github.com/uber/h3-js).
}
\author{
\strong{Maintainer}: Lauren O'Brien \email{obrlsoilau@gmail.com} (0000-0002-7336-2171)

}
