% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{point_to_h3}
\alias{point_to_h3}
\title{Convert point location to H3 address}
\usage{
point_to_h3(points = NULL, res = NULL, simple = TRUE)
}
\arguments{
\item{points}{object of class sfc_POINT. If crs is not WGS84, it will be
converted.}

\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{simple}{Logical; whether to return only outputs, or an `sf` object
containing both inputs and outputs.}
}
\value{
By default, a character vector of H3 addresses. if `simple = FALSE`,
  the input sf object is returned with new columns, one for each requested
  resolution. If `simple = TRUE` and multiple resolutions are requested, a
  data.frame is returned.
}
\description{
This function takes point location data and returns a H3 address for each
point at the chosen resolution(s).
}
\note{
While multiple resolutions can be requested for multiple points, be
  aware of the memory demand on large datasets.
}
\examples{
# where is the Brisbane Town Hall at resolution 15?
bth <- sf::st_sfc(sf::st_point(c(153.023503, -27.468920)), crs = 4326)
bth_15 <- point_to_h3(bth, res = 15)

# where is it at several resolutions?
bth_many <- point_to_h3(bth, res = seq(10, 15), simple = FALSE)
}
