% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_for_pt2h3}
\alias{prep_for_pt2h3}
\alias{prep_for_pt2h3.sf}
\alias{prep_for_pt2h3.sfc}
\alias{prep_for_pt2h3.sfg}
\alias{prep_for_pt2h3.matrix}
\alias{prep_for_pt2h3.data.frame}
\title{Prepare inputs for point_to_h3}
\usage{
prep_for_pt2h3(input = NULL)

\method{prep_for_pt2h3}{sf}(input = NULL)

\method{prep_for_pt2h3}{sfc}(input = NULL)

\method{prep_for_pt2h3}{sfg}(input = NULL)

\method{prep_for_pt2h3}{matrix}(input = NULL)

\method{prep_for_pt2h3}{data.frame}(input = NULL)
}
\arguments{
\item{input}{`sf`, `sfc` or `sfg` POLYGON/MULTIPOLYGON object, data frame or
matrix. Data frames or matrices must have x, y coordinates in their first
two columns.}
}
\value{
`sfc` representation of supplied coordinates. WGS84 input is assumed
  where input is non-spatial.
}
\description{
Sets up a variety of possible input objects for
\code{\link[h3jsr:point_to_h3]{h3jsr::point_to_h3()}}.
}
\keyword{internal}
