% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{res_cendist}
\alias{res_cendist}
\title{Get average distance between H3 points}
\usage{
res_cendist(res = NULL, units = c("m", "km"), fast = TRUE)
}
\arguments{
\item{res}{Integer; Desired H3 resolution. See
https://uber.github.io/h3/#/documentation/core-library/resolution-table for
allowable values and related dimensions.}

\item{units}{Length unit to report in, either meters or kilometers.}

\item{fast}{Logical; whether to retieve values from a locally stored table or
reclaculate from source.}
}
\value{
Numeric; h3 address center separation distance.
}
\description{
This function returns the average distance between the center of H3 address
hexagons at a given resolution.
}
\note{
This isn't in the core library but may be useful.
}
\examples{
# Return average h3 address separation distance at each resolution in kilometers
res_cendist(res = seq(0, 15), units = 'km')

}
