% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_utilities.R
\name{res_count}
\alias{res_count}
\title{Get total H3 cells}
\usage{
res_count(res = NULL, fast = TRUE)
}
\arguments{
\item{res}{Integer; Desired H3 resolution. See
\url{https://h3geo.org/docs/core-library/restable/} for allowable values and related dimensions.}

\item{fast}{Logical; whether to retieve values from a locally stored table or
reclaculate from source.}
}
\value{
Numeric; H3 cell count.
}
\description{
This function returns total number of H3 cells at a given resolution.
}
\examples{
# Return H3 cell count for resolution 8
res_count(res = 8)

}
